Reorganize repository to follow best practices
This change reorganizes the repository to follow standard Python project best practices.

- The main source code has been moved from `detectors/` to `src/detectors/`.
- All documentation files have been moved to a `docs/` directory.
- The `testing/` directory has been renamed to `tests/` and its contents reorganized.
- `pyproject.toml` has been updated to reflect the new structure, including paths for hatch, pytest, coverage, and ruff.
- The main `README.md` has been updated with the new project structure and file paths.
- The test scripts have been updated to work with the new structure, removing fragile `sys.path` hacks and fixing broken paths.
- Added `__init__.py` files to `tests/` and `tests/mock_server/` to make them proper packages.
