+++ README.md
@@ -17,10 +17,10 @@ This project provides enterprise-ready detection scripts to identify and assess 
 ### PowerShell Detector
 ```powershell
 # Basic scan
-.\detectors\detector.ps1 -TargetRange "192.168.1.1-50"
+.\src\detectors\detector.ps1 -TargetRange "192.168.1.1-50"
 
 # Scan with output file and verbose logging
-.\detectors\detector.ps1 -TargetRange "192.168.1.0/24" -OutputFile "results.json" -Verbose
+.\src\detectors\detector.ps1 -TargetRange "192.168.1.0/24" -OutputFile "results.json" -Verbose
 ```
 
 ### Python Detector
@@ -29,10 +29,10 @@ This project provides enterprise-ready detection scripts to identify and assess 
 uv sync
 
 # Basic scan
-uv run python detectors/detector.py --target-range "192.168.1.1-50"
+uv run python src/detectors/detector.py --target-range "192.168.1.1-50"
 
 # Scan with output file and verbose logging
-uv run python detectors/detector.py --target-range "10.0.0.0/24" --output results.json --verbose
+uv run python src/detectors/detector.py --target-range "10.0.0.0/24" --output results.json --verbose
 
 # Or use the installed script
 uv run cve-2025-53770-detect --target-range "192.168.1.1-50"
@@ -42,17 +42,28 @@ uv run cve-2025-53770-detect --target-range "192.168.1.1-50"
 
 ```
 CVE-2025-53770/
-├── README.md                    # Project documentation
-├── CLAUDE.md                    # Claude Code instructions
-├── NOTES.md                     # Reference links and resources
-├── TASKS.md                     # Project task tracking
-├── TODO.md                      # Prioritized backlog
-├── DETECTION_ALGORITHMS.md      # Pseudocode and algorithm documentation
-├── LIBRARIES.md                 # External dependencies documentation
-└── detectors/                   # Detection scripts
-    ├── detector.ps1             # PowerShell implementation
-    ├── detector.py              # Python implementation  
-    └── requirements.txt         # Python dependencies
+├── README.md
+├── docs/
+│   ├── CLAUDE.md
+│   ├── DETECTION_ALGORITHMS.md
+│   ├── LIBRARIES.md
+│   ├── NOTES.md
+│   ├── PROJ_NOTES.md
+│   ├── TASKS.md
+│   └── TODO.md
+├── src/
+│   └── detectors/
+│       ├── __init__.py
+│       ├── detector.py
+│       └── detector.ps1
+├── tests/
+│   ├── README.md
+│   ├── mock_server/
+│   ├── test_data/
+│   └── unit/
+├── .gitignore
+├── pyproject.toml
+└── uv.lock
 ```
 
 ## 🛠️ Installation & Setup
@@ -104,11 +115,11 @@ CVE-2025-53770/
 4. **Verify installation**
    ```bash
    # Test Python detector
-   uv run python detectors/detector.py --help
+   uv run python src/detectors/detector.py --help
    uv run cve-2025-53770-detect --help
    
    # Test PowerShell detector
-   Get-Help .\detectors\detector.ps1
+   Get-Help .\src\detectors\detector.ps1
    ```
 
 ## 📖 Usage Guide
@@ -126,13 +137,13 @@ CVE-2025-53770/
 #### Examples
 ```powershell
 # Scan single subnet with verbose output
-.\detector.ps1 -TargetRange "192.168.1.1-254" -Verbose
+.\src\detectors\detector.ps1 -TargetRange "192.168.1.1-254" -Verbose
 
 # Enterprise scan with results export
-.\detector.ps1 -TargetRange "10.0.0.1-100" -OutputFile "enterprise_scan.json"
+.\src\detectors\detector.ps1 -TargetRange "10.0.0.1-100" -OutputFile "enterprise_scan.json"
 ```
 
-### Python Detector (`detector.py`)
+### Python Detector (`src/detectors/detector.py`)
 
 #### Arguments
 - `--target-range, -t` (Required): IP range to scan
@@ -142,10 +153,10 @@ CVE-2025-53770/
 #### Examples
 ```bash
 # Comprehensive network scan
-uv run python detectors/detector.py --target-range "192.168.0.0/24" --verbose
+uv run python src/detectors/detector.py --target-range "192.168.0.0/24" --verbose
 
 # Targeted scan with results export
-uv run python detectors/detector.py -t "10.0.1.1-50" -o vulnerability_report.json
+uv run python src/detectors/detector.py -t "10.0.1.1-50" -o vulnerability_report.json
 
 # Using the installed command
 uv run cve-2025-53770-detect --target-range "192.168.0.0/24" --output scan_results.json
@@ -253,11 +264,11 @@ Both detectors generate structured JSON reports:
 
 ## 📚 Documentation
 
-- **[DETECTION_ALGORITHMS.md](DETECTION_ALGORITHMS.md)**: Detailed pseudocode and algorithm documentation
-- **[LIBRARIES.md](LIBRARIES.md)**: External dependencies and library documentation
-- **[TASKS.md](TASKS.md)**: Project development tasks and milestones
-- **[TODO.md](TODO.md)**: Prioritized development backlog
-- **[NOTES.md](NOTES.md)**: Reference links and resources
+- **[DETECTION_ALGORITHMS.md](docs/DETECTION_ALGORITHMS.md)**: Detailed pseudocode and algorithm documentation
+- **[LIBRARIES.md](docs/LIBRARIES.md)**: External dependencies and library documentation
+- **[TASKS.md](docs/TASKS.md)**: Project development tasks and milestones
+- **[TODO.md](docs/TODO.md)**: Prioritized development backlog
+- **[NOTES.md](docs/NOTES.md)**: Reference links and resources
 
 ## 🔗 References
