+++ docs/DETECTION_ALGORITHMS.md
