+++ docs/LIBRARIES.md
