+++ pyproject.toml
@@ -69,10 +69,10 @@ Repository = "https://github.com/example/cve-2025-53770"
 Issues = "https://github.com/example/cve-2025-53770/issues"
 
 [tool.hatch.build.targets.wheel]
-packages = ["detectors"]
+packages = ["src/detectors"]
 
 [tool.pytest.ini_options]
-testpaths = ["testing/unit-tests"]
+testpaths = ["tests/unit"]
 python_files = "test_*.py"
 python_classes = "Test*"
 python_functions = "test_*"
@@ -90,7 +90,7 @@ filterwarnings = [
 [tool.coverage.run]
 source = ["detectors"]
 omit = [
-    "testing/*",
+    "tests/*",
     "*/test_*",
 ]
 
@@ -145,7 +145,7 @@ ignore = [
 ]
 
 [tool.ruff.per-file-ignores]
-"testing/*" = ["B011"]
+"tests/*" = ["B011"]
 
 [tool.ruff.isort]
 known-first-party = ["detectors"]
