+++ src/detectors/detector.ps1
