+++ src/detectors/detector.py
@@ -205,6 +205,8 @@ class CVE202553770Detector:
         findings = []
         
         for scheme, data in probe_data.items():
+            if not isinstance(data, dict):
+                continue
             headers = data.get('headers', {})
             content = data.get('content', '')
             server = data.get('server', '')
