+++ tests/mock_server/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the mock_server directory a package.
