+++ tests/unit/run_tests.py
@@ -28,11 +28,12 @@ def run_command(cmd: list, description: str) -> bool:
     
     try:
         result = subprocess.run(cmd, capture_output=False, text=True)
-        if result.returncode == 0:
-            print(f"✅ {description} - PASSED")
+        # Pytest exit code 5 means no tests were collected, which is fine for skipped tests.
+        if result.returncode == 0 or result.returncode == 5:
+            print(f"✅ {description} - PASSED (or skipped)")
             return True
         else:
-            print(f"❌ {description} - FAILED")
+            print(f"❌ {description} - FAILED (exit code: {result.returncode})")
             return False
     except Exception as e:
         print(f"❌ {description} - ERROR: {e}")
