+++ tests/unit/test_powershell_detector.py
@@ -22,10 +22,8 @@ import threading
 from typing import Dict, Any, List
 from pathlib import Path
 
-# Import mock server for testing
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', 'mock-server'))
 try:
-    from mock_server import run_server
+    from tests.mock_server.mock_server import run_server
 except ImportError:
     pytest.skip("Mock server not available", allow_module_level=True)
 
@@ -37,7 +35,7 @@ class TestPowerShellDetector:
     def detector_script_path(self):
         """Get path to PowerShell detector script"""
         script_dir = Path(__file__).parent.parent.parent
-        script_path = script_dir / "detectors" / "detector.ps1"
+        script_path = script_dir / "src" / "detectors" / "detector.ps1"
         
         if not script_path.exists():
             pytest.skip("PowerShell detector script not found")
@@ -71,7 +69,7 @@ class TestPowerShellDetector:
         
         # Get detector script path
         script_dir = Path(__file__).parent.parent.parent
-        script_path = script_dir / "detectors" / "detector.ps1"
+        script_path = script_dir / "src" / "detectors" / "detector.ps1"
         
         if not script_path.exists():
             pytest.skip("PowerShell detector script not found")
@@ -256,7 +254,7 @@ class TestPowerShellDetectorPatterns:
         """Test that PowerShell detector covers expected patterns"""
         # Read the PowerShell script to verify patterns
         script_dir = Path(__file__).parent.parent.parent
-        script_path = script_dir / "detectors" / "detector.ps1"
+        script_path = script_dir / "src" / "detectors" / "detector.ps1"
         
         if not script_path.exists():
             pytest.skip("PowerShell detector script not found")
@@ -280,7 +278,7 @@ class TestPowerShellDetectorPatterns:
     def test_server_detection_patterns(self):
         """Test server detection patterns in PowerShell script"""
         script_dir = Path(__file__).parent.parent.parent
-        script_path = script_dir / "detectors" / "detector.ps1"
+        script_path = script_dir / "src" / "detectors" / "detector.ps1"
         
         if not script_path.exists():
             pytest.skip("PowerShell detector script not found")
