+++ tests/unit/test_python_detector.py
@@ -18,93 +18,88 @@ import os
 from unittest.mock import Mock, AsyncMock, patch, MagicMock
 from typing import Dict, Any, List
 
-# Add the detectors directory to the path
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..', 'detectors'))
-
 try:
-    from detector import CVE202553770Detector, VulnerabilityReport
+    from detectors.detector import CVE202553770Detector, VulnerabilityReport
 except ImportError as e:
     pytest.skip(f"Could not import detector module: {e}", allow_module_level=True)
 
 
-class TestCVE202553770Detector:
-    """Test class for the main detector functionality"""
-    
-    @pytest.fixture
-    def detector(self):
-        """Create a detector instance for testing"""
-        return CVE202553770Detector(verbose=False)
-    
-    @pytest.fixture
-    def high_risk_response(self):
-        """Mock high-risk HTTP response"""
-        return {
-            'http': {
-                'status_code': 200,
-                'headers': {
-                    'Server': 'Apache Tomcat/9.0.50',
-                    'Content-Type': 'text/html',
-                    'X-Java-Serialized-Object': 'true'
-                },
-                'content': '''
-                <html><body>
-                <h1>Enterprise Application</h1>
-                <p>This application uses ObjectInputStream for data processing.</p>
-                <p>Configuration: BinaryFormatter enabled, serialVersionUID validation disabled.</p>
-                <p>Warning: Direct deserialization of untrusted input detected.</p>
-                </body></html>
-                ''',
-                'server': 'Apache Tomcat/9.0.50',
-                'content_type': 'text/html',
-                'url': 'http://192.168.1.100:8080/'
-            }
+@pytest.fixture
+def detector():
+    """Create a detector instance for testing"""
+    return CVE202553770Detector(verbose=False)
+
+@pytest.fixture
+def high_risk_response():
+    """Mock high-risk HTTP response"""
+    return {
+        'http': {
+            'status_code': 200,
+            'headers': {
+                'Server': 'Apache Tomcat/9.0.50',
+                'Content-Type': 'text/html',
+                'X-Java-Serialized-Object': 'true'
+            },
+            'content': '''
+            <html><body>
+            <h1>Enterprise Application</h1>
+            <p>This application uses ObjectInputStream for data processing.</p>
+            <p>Configuration: BinaryFormatter enabled, serialVersionUID validation disabled.</p>
+            <p>Warning: Direct deserialization of untrusted input detected.</p>
+            </body></html>
+            ''',
+            'server': 'Apache Tomcat/9.0.50',
+            'content_type': 'text/html',
+            'url': 'http://192.168.1.100:8080/'
         }
-    
-    @pytest.fixture
-    def medium_risk_response(self):
-        """Mock medium-risk HTTP response"""
-        return {
-            'http': {
-                'status_code': 200,
-                'headers': {
-                    'Server': 'Python/3.9 Flask/2.0.1',
-                    'Content-Type': 'application/json'
-                },
-                'content': '''
-                {
-                    "app": "Python Web Service",
-                    "features": ["pickle.loads support", "yaml.load configuration"],
-                    "security": {"input_validation": false, "untrusted_sources": true}
-                }
-                ''',
-                'server': 'Python/3.9 Flask/2.0.1',
-                'content_type': 'application/json',
-                'url': 'http://192.168.1.101:8080/'
+    }
+
+@pytest.fixture
+def medium_risk_response():
+    """Mock medium-risk HTTP response"""
+    return {
+        'http': {
+            'status_code': 200,
+            'headers': {
+                'Server': 'Python/3.9 Flask/2.0.1',
+                'Content-Type': 'application/json'
+            },
+            'content': '''
+            {
+                "app": "Python Web Service",
+                "features": ["pickle.loads support", "yaml.load configuration"],
+                "security": {"input_validation": false, "untrusted_sources": true},
+                "notes": "This service performs unsafe deserialization."
             }
+            ''',
+            'server': 'Python/3.9 Flask/2.0.1',
+            'content_type': 'application/json',
+            'url': 'http://192.168.1.101:8080/'
         }
-    
-    @pytest.fixture
-    def clean_response(self):
-        """Mock clean HTTP response with no detection patterns"""
-        return {
-            'http': {
-                'status_code': 200,
-                'headers': {
-                    'Server': 'nginx/1.18.0',
-                    'Content-Type': 'text/html'
-                },
-                'content': '''
-                <html><body>
-                <h1>Secure Web Application</h1>
-                <p>This application follows security best practices.</p>
-                <p>All input is properly validated and sanitized.</p>
-                </body></html>
-                ''',
-                'server': 'nginx/1.18.0',
-                'content_type': 'text/html',
-                'url': 'http://192.168.1.102:80/'
-            }
+    }
+
+@pytest.fixture
+def clean_response():
+    """Mock clean HTTP response with no detection patterns"""
+    return {
+        'http': {
+            'status_code': 200,
+            'headers': {
+                'Server': 'nginx/1.18.0',
+                'Content-Type': 'text/html'
+            },
+            'content': '''
+            <html><body>
+            <h1>Secure Web Application</h1>
+            <p>This application follows security best practices.</p>
+            <p>All input is properly validated and sanitized.</p>
+            </body></html>
+            ''',
+            'server': 'nginx/1.18.0',
+            'content_type': 'text/html',
+            'url': 'http://192.168.1.102:80/'
         }
+    }
 
 
 class TestIPRangeParsing:
@@ -150,7 +145,8 @@ class TestPortScanning:
         with patch('asyncio.open_connection') as mock_conn:
             # Mock successful connection
             reader_mock = AsyncMock()
-            writer_mock = AsyncMock()
+            writer_mock = MagicMock()
+            writer_mock.wait_closed = AsyncMock()
             mock_conn.return_value = (reader_mock, writer_mock)
             
             result = await detector.scan_port("127.0.0.1", 80)
@@ -322,9 +318,9 @@ class TestHTTPProbing:
         mock_response = AsyncMock()
         mock_response.status = 200
         mock_response.headers = {'Server': 'nginx/1.18.0', 'Content-Type': 'text/html'}
-        mock_response.text.return_value = "<html><body>Test content</body></html>"
+        mock_response.text = AsyncMock(return_value="<html><body>Test content</body></html>")
         
-        mock_session = AsyncMock()
+        mock_session = MagicMock()
         mock_session.get.return_value.__aenter__.return_value = mock_response
         
         detector.session = mock_session
